this version is server used to client in websocket channel
