+++ lib/config/AppConfig.dart
@@ -17,9 +17,8 @@ import 'package:flutter/material.dart';
 import 'package:pretty_dio_logger/pretty_dio_logger.dart';
 import 'package:talker_dio_logger/talker_dio_logger.dart';
 import '../manager/ToolsManager.dart';
-import '../microService/chat/ChatList.dart';
-import '../microService/chat/ChatPage.dart';
-import '../microService/chat/PagePerson.dart';
+import '../microService/chat/pages/home/ChatList.dart';
+import '../microService/chat/pages/person/PagePerson.dart';
 import '../pages/mobile/Page2.dart';
 import '../pages/mobile/Page3.dart';
 import '../../widgets/BottomNavigationBars/animated_notch_bottom_bar.dart';
