+++ lib/manager/GlobalManager.dart
@@ -8,9 +8,12 @@
 import 'dart:convert';
 import 'package:app_template/microService/chat/websocket/model/MessageQueue.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter_chat_ui/flutter_chat_ui.dart' as types;
 import 'package:shared_preferences/shared_preferences.dart';
+
 //应用配置信息
 import '../database/LocalStorage.dart';
+import '../microService/chat/websocket/Client.dart';
 import '../microService/chat/websocket/model/ClientObject.dart';
 import '../microService/chat/websocket/model/OffLineMessageQueue.dart';
 import '../microService/chat/websocket/schedule/UserSchedule.dart';
@@ -45,6 +48,10 @@ class GlobalManager {
   static late BuildContext context;
   // 8.全局add user消息队列
   static MessageQueue offerUserAddQueue = MessageQueue();
+  // 9.全局聊天消息存储
+  static List<types.Message> GloablMessageList = [];
+  // 10.全局chatWebsocketClient
+  static ChatWebsocketClient? chatWebsocketClient;
 
   /**************↑↑↑↑↑↑↑↑全局参数变量初始化操作↑↑↑↑↑↑↑↑***************/
 
@@ -84,6 +91,9 @@ class GlobalManager {
     // 调试管理器模块
     TestManager.debug();
 
+    // 设置deviceId
+    // String deviceId = await UniqueDeviceId.getDeviceUuid();
+    // appCache.setString("deviceId", deviceId);
     // 监测app是否初次启动
     final prefs = await SharedPreferences.getInstance();
     final isFirstRun = prefs.getBool('isFirstRun') ?? true;
