+++ lib/microService/chat/pages/chat/model/ServerTranslateClientChannel.dart
@@ -1,0 +1,8 @@
+/*
+ 这是server端转client接收端的类
+ 详情: 当server作为client需要与其他client进行通讯时主要运用该类
+ */
+
+class ServerTranslateClientChannel {
+  //
+}
