+++ lib/microService/chat/pages/home/ChatList.dart
@@ -6,8 +6,8 @@ import 'package:flutter/material.dart';
 import 'package:flutter_slidable/flutter_slidable.dart';
 import 'package:popup_menu_plus/popup_menu_plus.dart';
 import 'package:pull_to_refresh/pull_to_refresh.dart';
-import '../../config/AppConfig.dart';
-import '../../database/daos/UserDao.dart';
+import '../../../../config/AppConfig.dart';
+import '../../../../database/daos/UserDao.dart';
 import 'package:random_avatar/random_avatar.dart';
 
 class ChatList extends StatefulWidget {
@@ -21,6 +21,7 @@ class _ChatListState extends State<ChatList> {
   _ChatListState() {
     userList = [];
   }
+  UserDao userDao = UserDao();
   GlobalKey btnKeyMenu = GlobalKey();
   PopupMenu? menu;
   int page = 1; //页码
@@ -34,7 +35,6 @@ class _ChatListState extends State<ChatList> {
   @override
   void initState() {
     super.initState();
-    UserDao userDao = UserDao();
 
     // 加载websocket
     GlobalManager().GlobalChatWebsocket.bootWebsocket();
@@ -52,81 +52,62 @@ class _ChatListState extends State<ChatList> {
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      appBar: AppBar(
-        backgroundColor: Colors.grey,
-        iconTheme: IconThemeData(
-          color: Colors.blue,
-          opacity: 0.5,
-        ),
-        leadingWidth: 50,
-        titleSpacing: 1,
-        leading: Builder(builder: (BuildContext context) {
-          return IconButton(
-            iconSize: 25,
-            icon: const Icon(
-              Icons.chat_bubble_outlined,
-              color: Colors.blue,
+        appBar: AppBar(
+          backgroundColor: Colors.grey,
+          iconTheme: IconThemeData(
+            color: Colors.blue,
+            opacity: 0.5,
+          ),
+          leadingWidth: 50,
+          titleSpacing: 1,
+          leading: Builder(builder: (BuildContext context) {
+            return IconButton(
+              iconSize: 25,
+              icon: const Icon(
+                Icons.chat_bubble_outlined,
+                color: Colors.blue,
+              ),
+              onPressed: () {
+                //
+              },
+            );
+          }),
+          title: Column(children: [
+            Text(
+              AppConfig.appConfig['name'].toString().tr(),
+              style: const TextStyle(fontSize: 18),
             ),
-            onPressed: () {
-              //
-            },
-          );
-        }),
-        title: Column(children: [
-          Text(
-            AppConfig.appConfig['name'].toString().tr(),
-            style: const TextStyle(fontSize: 18),
-          ),
-          Text(
-            "我是小标题".tr(),
-            style: const TextStyle(fontSize: 10),
-          ),
-        ]),
-        actions: [
-          IconButton(
-            key: btnKeyMenu,
-            onPressed: () {
-              showMenu();
-            },
-            icon: const Icon(
-              Icons.add,
-              size: 35,
+            Text(
+              "我是小标题".tr(),
+              style: const TextStyle(fontSize: 10),
             ),
-          )
-        ],
-        actionsIconTheme: IconThemeData(
-          color: Colors.blue,
-        ),
-        shadowColor: Theme.of(context).shadowColor,
-        flexibleSpace: SizedBox(
-          width: double.infinity,
-          height: 160,
-          child: Image.network(
-            "https://ts1.cn.mm.bing.net/th/id/R-C.d4822697ad0424efafe6b62e5e6e0d1d?rik=ZdcMlu%2f2ng6ltA&riu=http%3a%2f%2fimg95.699pic.com%2fphoto%2f40141%2f5356.gif_wh860.gif&ehk=OMCk8kp7dU8UKPdcHORkcrjitRqABE0xoh7sa%2baGN4k%3d&risl=1&pid=ImgRaw&r=0",
-            fit: BoxFit.cover,
+          ]),
+          actions: [
+            IconButton(
+              key: btnKeyMenu,
+              onPressed: () {
+                showMenu();
+              },
+              icon: const Icon(
+                Icons.add,
+                size: 35,
+              ),
+            )
+          ],
+          actionsIconTheme: IconThemeData(
+            color: Colors.blue,
+          ),
+          shadowColor: Theme.of(context).shadowColor,
+          flexibleSpace: SizedBox(
+            width: double.infinity,
+            height: 160,
+            child: Image.network(
+              "https://ts1.cn.mm.bing.net/th/id/R-C.d4822697ad0424efafe6b62e5e6e0d1d?rik=ZdcMlu%2f2ng6ltA&riu=http%3a%2f%2fimg95.699pic.com%2fphoto%2f40141%2f5356.gif_wh860.gif&ehk=OMCk8kp7dU8UKPdcHORkcrjitRqABE0xoh7sa%2baGN4k%3d&risl=1&pid=ImgRaw&r=0",
+              fit: BoxFit.cover,
+            ),
           ),
         ),
-      ),
-      body: SmartRefresher(
-        enablePullDown: true,
-        enablePullUp: true,
-        header: const WaterDropHeader(),
-        controller: _refreshController,
-        onRefresh: _onRefresh,
-        onLoading: _onLoading,
-        child: listUser(),
-      ),
-    );
-  }
-
-  void _onRefresh() async {
-    await Future.delayed(const Duration(milliseconds: 1000));
-    _refreshController.refreshCompleted();
-  }
-
-  void _onLoading() async {
-    await Future.delayed(const Duration(milliseconds: 1000));
-    _refreshController.loadComplete();
+        body: listUser());
   }
 
   Widget listUser() {
@@ -217,9 +198,5 @@ class _ChatListState extends State<ChatList> {
       onDismiss: onDismiss,
     );
     menu.show(widgetKey: btnKeyMenu);
-  }
-
-  void onDismiss() {
-    print('Menu is dismissed');
   }
 }
